requirements.txt
httpx==0.19.0
Pillow==8.0.0
opencv-python==4.8.1.78
matplotlib==3.8.1
numpy==1.26.1
aiofiles==0.7.0 
flask==2.1.1
Werkzeug==2.1.1
flask-cors==4.0.0
aiohttp==3.9.0b1
